refactor: move JobStatus to its own file
